Updated the way the help button works
Reverted to using the default Windows help control box button. This simplifies things a lot.
We lose the ability to display a help button in the control box when there is a minimize
and/or maximize button visible, but in these cases, such forms should have either a menu that
includes a help option or a normal button to open help.

+ Removed the dependency on OpenDentalHelp.dll as this was a closed source and obfuscated
  assembly. Similar functionality will be implemented into the main program in the future.
  For the moment the help button does nothing.